Expand star rating system from 1-5 to 0.5-8 with half-star support
Major rating system improvements:
- Expanded rating range from 1-5 stars to 0.5-8 stars for more granular tracking
- Added half-star support (🌟 0.5) for minimal activities
- Updated keyboard layout with 3 rows: [0.5,1,2], [3,4,5], [6,7,8] + "Not done"
- Enhanced visual star display with proper formatting for decimals

Technical changes:
- Modified score keyboard in keyboards.py to show all new rating options
- Updated handlers.py to handle float values instead of int for star ratings
- Enhanced messages.py with format_stars_display() function and improved formatting
- Updated Google Sheets integration to properly handle decimal values (float parsing)
- Improved reminder statistics to display new star format correctly
- Updated help text and documentation to reflect 0.5-8 rating system

User benefits:
- More precise activity tracking (e.g., 0.5 for brief activities, 6-8 for exceptional days)
- Better granularity for distinguishing between different activity levels
- Maintained backward compatibility with existing data
- Cleaner display formatting in all interfaces (bot, reminders, stats)

+++ src/bot/handlers.py
@@ -180,7 +180,7 @@ async def handle_stars_selection(query, user_id: int, data: str):
     parts = data.split(":")
     if len(parts) >= 4:
         hobby_key = parts[1]
-        stars = int(parts[2])
+        stars = float(parts[2])  # Теперь поддерживаем десятичные значения
         target_date = parts[3]
         
         # Сохраняем увлечение в историю
